Bai2/requirements.txt
asgiref==3.8.1
Django==4.1.13
djongo==1.3.6
dnspython==2.6.1
pymongo==4.6.3
pytz==2024.1
sqlparse==0.2.4
typing_extensions==4.10.0
tzdata==2024.1
djangocorsheaders==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.1.2
django-environ==0.8.0